+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - **breaking change** the `CODE_SEGMENT_START` is now equal to 4
 - fixing a bug in the bytecode reader: it didn't handle `NOP`
 - `import` should be able to load plugins, also `import` takes only one argument now
+- **breaking change** `POP_JUMP_IF_FALSE` is now an absolute jump as well
 
 ### Removed
 - `hastype` keyword because I never had to implement compile time typechecking, so it's not useful